requirements.txt
streamlit
faiss-cpu
pandas
sentence-transformers
openai
langchain
transformers
bitsandbytes
